Tweaked for server, sub, and pub on Ubuntu.
It didn't build for me out of the box, I had to add `-lrt` to the
binaries. Whether or not that also works on Mac OS is another
question. I'd be happy to adjust the PR as needed.

I also added copy-pastable build and run commands. I hadn't used
`bazel` in a while and figure it might help others to have them handy.